Change service monitoring to runtime
